requirements.txt
annotated-types==0.7.0
anyio==4.12.0
attrs==25.4.0
autoevals==0.0.130
boto3==1.42.18
botocore==1.42.18
braintrust==0.3.15
certifi==2025.11.12
charset-normalizer==3.4.4
chevron==0.14.0
click==8.3.1
distro==1.9.0
exceptiongroup==1.3.1
gitdb==4.0.12
GitPython==3.1.45
googleapis-common-protos==1.72.0
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.11
importlib_metadata==8.7.1
jiter==0.12.0
jmespath==1.0.1
jsonschema==4.25.1
jsonschema-specifications==2025.9.1
openai==2.14.0
opentelemetry-api==1.39.1
opentelemetry-exporter-otlp-proto-common==1.39.1
opentelemetry-exporter-otlp-proto-http==1.39.1
opentelemetry-proto==1.39.1
opentelemetry-sdk==1.39.1
opentelemetry-semantic-conventions==0.60b1
polyleven==0.9.0
protobuf==6.33.2
psycopg2-binary==2.9.11
pydantic==2.12.5
pydantic_core==2.41.5
python-dateutil==2.9.0.post0
python-dotenv==1.2.1
python-slugify==8.0.4
PyYAML==6.0.3
referencing==0.37.0
requests==2.32.5
rpds-py==0.30.0
s3transfer==0.16.0
six==1.17.0
smmap==5.0.2
sniffio==1.3.1
sseclient-py==1.8.0
starlette==0.50.0
text-unidecode==1.3
tqdm==4.67.1
typing-inspection==0.4.2
typing_extensions==4.15.0
urllib3==2.6.2
uv==0.9.20
uvicorn==0.40.0
wrapt==2.0.1
zipp==3.23.0
